lesson 61

+++ src/main/java/kg/attractor/movie_review_21/service/DirectorService.java
@@ -9,7 +9,7 @@ public interface DirectorService {
 
     List<DirectorDto> getList();
 
-    DirectorDto getByMovieId(int movieId);
+    DirectorDto getByMovieId(long movieId);
 
     void create(DirectorDto directorDto);
 }
